change cache clearing

+++ rakaia/app.py
@@ -6,7 +6,7 @@ from dash_extensions.enrich import DashProxy, ServersideOutputTransform, FileSys
 import dash_bootstrap_components as dbc
 from rakaia.components.layout import register_app_layout
 from rakaia.callbacks.pixel import init_pixel_level_callbacks
-from rakaia.callbacks.object import init_cell_level_callbacks
+from rakaia.callbacks.object import init_object_level_callbacks
 from rakaia.callbacks.roi import init_roi_level_callbacks
 from rakaia.callbacks.db import init_db_callbacks
 
@@ -21,12 +21,14 @@ def init_dashboard(server, authentic_id, config=None):
     with tempfile.TemporaryDirectory() as tmpdirname:
         # set the server output cache dir and clean it every time a new dash session is started
         # if whatever reason, the tmp is not writable, use a new directory as a backup
-        if os.access(config['cache_dest'], os.R_OK):
-            # cleaning the tmp dir for any sub directory that has rakaia cache in it
+        # do not try to clear the caches if not using overwrite, for example deployed to a shared server
+        if os.access(config['cache_dest'], os.R_OK) and config['serverside_overwrite']:
+            # cleaning the tmp dir for any sub-directory that has rakaia cache in it
             cache_subdirs = [x[0] for x in os.walk(config['cache_dest']) if 'rakaia_cache' in x[0]]
             # remove any parent directory that has a rakaia cache in it
             for cache_dir in cache_subdirs:
-                if os.access(os.path.dirname(cache_dir), os.R_OK) and os.access(cache_dir, os.R_OK):
+                if os.access(os.path.dirname(cache_dir), os.R_OK) and os.access(cache_dir, os.R_OK) and \
+                        os.path.isdir(cache_dir):
                     shutil.rmtree(os.path.dirname(cache_dir))
             cache_dest = os.path.join(config['cache_dest'], authentic_id, "rakaia_cache")
         else:
@@ -90,7 +92,7 @@ def init_dashboard(server, authentic_id, config=None):
     dash_app.enable_dev_tools(debug=config['is_dev_mode'])
 
     init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, config)
-    init_cell_level_callbacks(dash_app, tmpdirname, authentic_id, config)
+    init_object_level_callbacks(dash_app, tmpdirname, authentic_id, config)
     init_roi_level_callbacks(dash_app, tmpdirname, authentic_id, config)
     init_db_callbacks(dash_app, tmpdirname, authentic_id, config)
